fix:修复服务bug

+++ pay/internal/repository/pay.go
@@ -3,10 +3,10 @@ package repository
 import "gorm.io/gorm"
 
 type Pay struct {
-	PayID  int     `gorm:"primaryKey" json:"id"`
-	UserID int     `json:"userId"`
-	ItemID int     `json:"itemId"`
-	Amount float64 `json:"amount"`
+	PayID  int `gorm:"primaryKey" json:"id"`
+	UserID int `json:"userId"`
+	ItemID int `json:"itemId"`
+	Amount int `json:"amount"`
 }
 
 type PayRepository interface {
@@ -21,7 +21,7 @@ func NewPayRepositoryImpl(db *gorm.DB) *PayRepositoryImpl {
 	return &PayRepositoryImpl{db: db}
 }
 
-func (p *PayRepositoryImpl) CreatePay(userID int, itemID int, amount float64) error {
+func (p *PayRepositoryImpl) CreatePay(userID int, itemID int, amount int) error {
 	pay := &Pay{
 		UserID: userID,
 		ItemID: itemID,
